refactor: refactored dbot tours
