+++ packages/bot-web-ui/src/components/dashboard/dashboard.tsx
@@ -11,13 +11,12 @@ import { DBOT_TABS, TAB_IDS } from 'Constants/bot-contents';
 import { useDBotStore } from 'Stores/useDBotStore';
 import RunPanel from '../run-panel';
 import RunStrategy from './dashboard-component/run-strategy';
-import { DBOT_ONBOARDING } from './dbot-tours/config';
+import TourTriggrerDialog from './dbot-tours/common/tour-trigger-dialog';
+import DesktopTours from './dbot-tours/desktop-tours';
 import MobileTours from './dbot-tours/mobile-tours';
 import { getTourSettings, setTourSettings, setTourType, tour_status_ended, tour_type } from './dbot-tours/utils';
 import DashboardComponent from './dashboard-component';
-import ReactJoyrideWrapper from './react-joyride-wrapper';
 import StrategyNotification from './strategy-notification';
-import TourTriggrerDialog from './tour-trigger-dialog';
 import Tutorial from './tutorial-tab';
 
 const Dashboard = observer(() => {
@@ -225,13 +224,7 @@ const Dashboard = observer(() => {
                 >
                     {(active_tab === DASHBOARD || active_tab === BOT_BUILDER) && <TourTriggrerDialog />}
 
-                    {has_tour_started &&
-                        active_tab === DASHBOARD &&
-                        (is_mobile ? (
-                            <MobileTours />
-                        ) : (
-                            <ReactJoyrideWrapper steps={DBOT_ONBOARDING} spotlightClicks hideCloseButton />
-                        ))}
+                    {has_tour_started && active_tab === DASHBOARD && (is_mobile ? <MobileTours /> : <DesktopTours />)}
                     <Tabs
                         active_index={active_tab}
                         className='dashboard__tabs'
