+++ packages/bot-web-ui/src/components/dashboard/dbot-tours/desktop-tours/common/react-joyride-wrapper.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactJoyride, { Step, Styles } from 'react-joyride';
 import { localize } from '@deriv/translations';
-import { handleJoyrideCallback } from './dbot-tours/utils';
+import { handleJoyrideCallback } from '../../utils/index';
 
 const common_tour_button_properties = {
     fontWeight: '700',
