+++ packages/bot-web-ui/src/components/dashboard/dbot-tours/desktop-tours/index.ts
@@ -1,0 +1,3 @@
+import DesktopTours from './desktop-tours';
+
+export default DesktopTours;
